backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
jinja2==3.1.2
weasyprint==60.1
requests==2.31.0
pydantic==2.5.0
tzdata==2026.1
apscheduler==3.10.4
